Refactor tests - second test suite for apexRootDirectory
